+++ examples/nvexec/maxwell_distributed.cpp
@@ -21,7 +21,7 @@
 #include <vector>
 
 static std::pair<std::size_t, std::size_t>
-  even_share(std::size_t n, std::uint32_t rank, std::uint32_t size) noexcept {
+  even_share(std::size_t n, std::size_t rank, std::size_t size) noexcept {
   const auto avg_per_thread = n / size;
   const auto n_big_share = avg_per_thread + 1;
   const auto big_shares = n % size;
